requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
chromadb>=0.4.22
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pytest>=8.0.0
httpx>=0.26.0
